Finally adding fish in fish grabbing game after 5 days
me sigma me do fish and catch fish logic

+++ Assets/CashScript.cs
@@ -6,7 +6,7 @@ using UnityEngine.UI;
 
 public class CashScript : MonoBehaviour
 {
-    public double playerCash;
+    public float playerCash;
     public Text cashText;
 
     private void Update()
